Move all legacy code into operations/legacy
