+++ mcp/src/main/java/io/airlift/mcp/McpModule.java
@@ -38,10 +38,13 @@ import io.airlift.mcp.model.JsonRpcMessageDeserializer;
 import io.airlift.mcp.model.JsonSchemaBuilder;
 import io.airlift.mcp.model.JsonSchemaBuilder.DefaultSchemaBuilderProvider;
 import io.airlift.mcp.model.JsonSchemaBuilder.SchemaBuilder;
-import io.airlift.mcp.operations.LegacyOperations;
 import io.airlift.mcp.operations.Operations;
 import io.airlift.mcp.operations.OperationsModule;
-import io.airlift.mcp.operations.SessionlessOperations;
+import io.airlift.mcp.operations.legacy.LegacyOperations;
+import io.airlift.mcp.operations.legacy.SessionlessOperations;
+import io.airlift.mcp.operations.legacy.sessions.CachingSessionController;
+import io.airlift.mcp.operations.legacy.sessions.ForSessionCaching;
+import io.airlift.mcp.operations.legacy.sessions.SessionController;
 import io.airlift.mcp.reflection.AppContent;
 import io.airlift.mcp.reflection.CompletionHandlerProvider;
 import io.airlift.mcp.reflection.IconHelper;
@@ -51,9 +54,6 @@ import io.airlift.mcp.reflection.ReflectionHelper;
 import io.airlift.mcp.reflection.ResourceHandlerProvider;
 import io.airlift.mcp.reflection.ResourceTemplateHandlerProvider;
 import io.airlift.mcp.reflection.ToolHandlerProvider;
-import io.airlift.mcp.sessions.CachingSessionController;
-import io.airlift.mcp.sessions.ForSessionCaching;
-import io.airlift.mcp.sessions.SessionController;
 import io.airlift.mcp.storage.StorageController;
 
 import java.lang.annotation.Annotation;
