+++ mcp/src/main/java/io/airlift/mcp/operations/PaginationUtil.java
@@ -12,16 +12,16 @@ import java.util.stream.Stream;
 import static com.google.common.collect.ImmutableList.toImmutableList;
 import static java.util.Comparator.comparing;
 
-class PaginationUtil
+public class PaginationUtil
 {
     private final int pageSize;
 
-    PaginationUtil(McpConfig mcpConfig)
+    public PaginationUtil(McpConfig mcpConfig)
     {
         pageSize = mcpConfig.getDefaultPageSize();
     }
 
-    <T, L> T paginate(PaginatedRequest request, List<L> items, Function<L, String> keyMapper, BiFunction<List<L>, Optional<String>, T> resultMapper)
+    public <T, L> T paginate(PaginatedRequest request, List<L> items, Function<L, String> keyMapper, BiFunction<List<L>, Optional<String>, T> resultMapper)
     {
         Stream<L> itemsStream = request.cursor()
                 .map(cursor -> items.stream()
