+++ mcp/src/main/java/io/airlift/mcp/operations/legacy/LegacyOperations.java
@@ -1,4 +1,4 @@
-package io.airlift.mcp.operations;
+package io.airlift.mcp.operations.legacy;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.json.JsonMapper;
@@ -41,10 +41,11 @@ import io.airlift.mcp.model.SetLevelRequest;
 import io.airlift.mcp.model.SubscribeListChanged;
 import io.airlift.mcp.model.SubscribeRequest;
 import io.airlift.mcp.model.Tool;
+import io.airlift.mcp.operations.Operations;
+import io.airlift.mcp.operations.legacy.sessions.SessionController;
+import io.airlift.mcp.operations.legacy.sessions.SessionId;
+import io.airlift.mcp.operations.legacy.sessions.SessionValueKey;
 import io.airlift.mcp.reflection.IconHelper;
-import io.airlift.mcp.sessions.SessionController;
-import io.airlift.mcp.sessions.SessionId;
-import io.airlift.mcp.sessions.SessionValueKey;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.ws.rs.WebApplicationException;
@@ -82,14 +83,14 @@ import static io.airlift.mcp.model.Constants.NOTIFICATION_ROOTS_LIST_CHANGED;
 import static io.airlift.mcp.model.JsonRpcErrorCode.INVALID_PARAMS;
 import static io.airlift.mcp.model.JsonRpcErrorCode.METHOD_NOT_FOUND;
 import static io.airlift.mcp.model.Protocol.LATEST_PROTOCOL;
-import static io.airlift.mcp.operations.OperationsCommon.supportsIcons;
-import static io.airlift.mcp.sessions.SessionValueKey.CLIENT_CAPABILITIES;
-import static io.airlift.mcp.sessions.SessionValueKey.LOGGING_LEVEL;
-import static io.airlift.mcp.sessions.SessionValueKey.PROTOCOL;
-import static io.airlift.mcp.sessions.SessionValueKey.ROOTS;
-import static io.airlift.mcp.sessions.SessionValueKey.SENT_MESSAGES;
-import static io.airlift.mcp.sessions.SessionValueKey.cancellationKey;
-import static io.airlift.mcp.sessions.SessionValueKey.serverToClientResponseKey;
+import static io.airlift.mcp.operations.legacy.OperationsCommon.supportsIcons;
+import static io.airlift.mcp.operations.legacy.sessions.SessionValueKey.CLIENT_CAPABILITIES;
+import static io.airlift.mcp.operations.legacy.sessions.SessionValueKey.LOGGING_LEVEL;
+import static io.airlift.mcp.operations.legacy.sessions.SessionValueKey.PROTOCOL;
+import static io.airlift.mcp.operations.legacy.sessions.SessionValueKey.ROOTS;
+import static io.airlift.mcp.operations.legacy.sessions.SessionValueKey.SENT_MESSAGES;
+import static io.airlift.mcp.operations.legacy.sessions.SessionValueKey.cancellationKey;
+import static io.airlift.mcp.operations.legacy.sessions.SessionValueKey.serverToClientResponseKey;
 import static io.opentelemetry.semconv.incubating.McpIncubatingAttributes.MCP_METHOD_NAME;
 import static io.opentelemetry.semconv.incubating.McpIncubatingAttributes.MCP_PROTOCOL_VERSION;
 import static io.opentelemetry.semconv.incubating.McpIncubatingAttributes.MCP_RESOURCE_URI;
