+++ mcp/src/main/java/io/airlift/mcp/operations/legacy/ResourceVersions.java
@@ -1,4 +1,4 @@
-package io.airlift.mcp.operations;
+package io.airlift.mcp.operations.legacy;
 
 import com.google.common.collect.ImmutableMap;
