+++ mcp/src/main/java/io/airlift/mcp/operations/legacy/SystemListVersions.java
@@ -1,4 +1,4 @@
-package io.airlift.mcp.operations;
+package io.airlift.mcp.operations.legacy;
 
 import static java.util.Objects.requireNonNull;
