+++ mcp/src/main/java/io/airlift/mcp/operations/legacy/sessions/ForSessionCaching.java
@@ -1,4 +1,4 @@
-package io.airlift.mcp.sessions;
+package io.airlift.mcp.operations.legacy.sessions;
 
 import com.google.inject.BindingAnnotation;
