+++ mcp/src/main/java/io/airlift/mcp/operations/legacy/sessions/SessionId.java
@@ -1,4 +1,4 @@
-package io.airlift.mcp.sessions;
+package io.airlift.mcp.operations.legacy.sessions;
 
 import com.fasterxml.jackson.annotation.JsonValue;
