+++ mcp/src/test/java/io/airlift/mcp/TestConformance.java
@@ -4,7 +4,7 @@ import com.google.common.collect.ImmutableMap;
 import com.google.common.io.Closer;
 import com.google.inject.Scopes;
 import io.airlift.http.server.testing.TestingHttpServer;
-import io.airlift.mcp.sessions.StandardSessionController;
+import io.airlift.mcp.operations.legacy.sessions.StandardSessionController;
 import io.airlift.mcp.storage.MemoryStorageController;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.TestInstance;
