+++ mcp/src/test/java/io/airlift/mcp/TestDatabaseSessionController.java
@@ -2,8 +2,8 @@ package io.airlift.mcp;
 
 import com.fasterxml.jackson.databind.json.JsonMapper;
 import io.airlift.json.JsonMapperProvider;
-import io.airlift.mcp.sessions.SessionController;
-import io.airlift.mcp.sessions.StandardSessionController;
+import io.airlift.mcp.operations.legacy.sessions.SessionController;
+import io.airlift.mcp.operations.legacy.sessions.StandardSessionController;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.TestInstance;
