+++ mcp/src/test/java/io/airlift/mcp/TestMemorySessionController.java
@@ -2,10 +2,10 @@ package io.airlift.mcp;
 
 import io.airlift.json.JsonMapperProvider;
 import io.airlift.mcp.McpIdentity.Authenticated;
-import io.airlift.mcp.sessions.SessionController;
-import io.airlift.mcp.sessions.SessionId;
-import io.airlift.mcp.sessions.SessionValueKey;
-import io.airlift.mcp.sessions.StandardSessionController;
+import io.airlift.mcp.operations.legacy.sessions.SessionController;
+import io.airlift.mcp.operations.legacy.sessions.SessionId;
+import io.airlift.mcp.operations.legacy.sessions.SessionValueKey;
+import io.airlift.mcp.operations.legacy.sessions.StandardSessionController;
 import io.airlift.mcp.storage.MemoryStorageController;
 
 import java.time.Duration;
