+++ mcp/src/test/java/io/airlift/mcp/TestSentMessages.java
@@ -11,10 +11,10 @@ import io.airlift.http.server.testing.TestingHttpServer;
 import io.airlift.mcp.messages.SentMessages;
 import io.airlift.mcp.messages.SentMessages.SentMessage;
 import io.airlift.mcp.model.Icon;
-import io.airlift.mcp.sessions.ForSessionCaching;
-import io.airlift.mcp.sessions.SessionController;
-import io.airlift.mcp.sessions.SessionId;
-import io.airlift.mcp.sessions.StandardSessionController;
+import io.airlift.mcp.operations.legacy.sessions.ForSessionCaching;
+import io.airlift.mcp.operations.legacy.sessions.SessionController;
+import io.airlift.mcp.operations.legacy.sessions.SessionId;
+import io.airlift.mcp.operations.legacy.sessions.StandardSessionController;
 import io.airlift.mcp.storage.MemoryStorageController;
 import io.modelcontextprotocol.spec.McpSchema.CallToolRequest;
 import org.junit.jupiter.api.AfterAll;
